3 files: V Story: extrace travel story
